Add files via upload
A URL shortener implemented in Python is a utility that simplifies lengthy URLs while maintaining their functionality. It achieves this by generating a unique alphanumeric code for each original URL. When a user inputs a long URL, the shortening script generates this code, which is appended to a shortened domain (e.g., "short.url/abc123"). Upon clicking the shortened link, the server extracts the code, maps it to the corresponding original URL, and redirects the user. This process enhances link sharing, minimizes character count, and offers tracking capabilities like click analytics. It involves string manipulation, data storage, and redirection mechanisms to efficiently manage and redirect both short and original URLs.